Make OkHttpClient implement both Client and AsyncClient
Removes the need to share code in an abstract class.
